Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pillow==10.4.0
python-dotenv==1.0.0